Major suggestions to the onepopperiod.rmd
I made a lot of wordsmithing suggestions, and suggest only very minor edits to the R code chunks.

The biggest suggestion was the organization -- I tried, and perhaps unsuccessfully -- to split the vignette into hierarchies: Sections and Steps. Only because I was kind of overwhelmed with what was going on in the vignette.

I was tempted to echo==FALSE in chunk options, because that is also very distracting to me, howeve,r, I know that in some chunks an echo would be good...